# public/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "OCPD Insurance Platform API",
        "description": "REST API dla platformy ubezpieczeń odpowiedzialności cywilnej przewoźnika drogowego (OCPD).\n\n## Funkcjonalności\n- Zarządzanie klientami (przewoźnikami)\n- Zarządzanie polisami OCPD\n- Kalkulator składek z oceną ryzyka\n- Obsługa szkód\n- Statystyki dashboardu",
        "version": "1.0.0",
        "contact": {
            "name": "OCPD Support",
            "email": "support@ocpd-platform.pl"
        }
    },
    "servers": [
        {
            "url": "http://localhost:3000/api",
            "description": "Development server"
        }
    ],
...
    "paths": {
...
        "/policies": {
            "get": {
                "tags": [
                    "Policies"
                ],
                "summary": "Lista polis",
                "operationId": "getPolicies",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "DRAFT",
                                "QUOTED",
                                "ACTIVE",
                                "EXPIRED",
                                "CANCELLED"
                            ]
                        }
                    },
                    {
                        "name": "scope",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "POLAND",
                                "EUROPE",
                                "WORLD"
                            ]
                        }
                    },
                    {
                        "name": "clientId",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista polis ze statystykami",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Policy"
                                            }
                                        },
                                        "total": {
                                            "type": "integer"
                                        },
                                        "stats": {
                                            "$ref": "#/components/schemas/PolicyStats"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Policies"
                ],
                "summary": "Utwórz nową polisę",
                "operationId": "createPolicy",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PolicyInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Polisa utworzona",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Policy"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    }
                }
            }
        },
        "/policies/{id}": {
            "get": {
                "tags": [
                    "Policies"
                ],
                "summary": "Pobierz polisę",
                "operationId": "getPolicy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dane polisy",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Policy"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Policies"
                ],
                "summary": "Zaktualizuj status polisy",
                "operationId": "updatePolicyStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "DRAFT",
                                            "QUOTED",
                                            "ACTIVE",
                                            "EXPIRED",
                                            "CANCELLED"
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status zaktualizowany"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            }
        },
...
        "/quotes/calculate": {
            "post": {
                "tags": [
                    "Quotes"
                ],
                "summary": "Kalkulacja składki",
                "description": "Oblicza składkę ubezpieczeniową na podstawie parametrów. Obsługuje dwa tryby:\n- **Quick Quote** - szybka wycena z zakresem min/max\n- **Full Quote** - pełna kalkulacja z oceną ryzyka i referral",
                "operationId": "calculatePremium",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CalculationInput"
                            },
                            "examples": {
                                "quickQuote": {
                                    "summary": "Szybka wycena",
                                    "value": {
                                        "sumInsured": 300000,
                                        "territorialScope": "EUROPE",
                                        "quickQuote": true
                                    }
                                },
                                "fullQuote": {
                                    "summary": "Pełna kalkulacja",
                                    "value": {
                                        "sumInsured": 500000,
                                        "territorialScope": "EUROPE",
                                        "selectedClauses": [
                                            "GROSS_NEGLIGENCE",
                                            "FRIDGE"
                                        ],
                                        "yearsInBusiness": 10,
                                        "fleetSize": 15,
                                        "apkData": {
                                            "mainCargoTypes": [
                                                "Elektronika"
                                            ],
                                            "averageCargoValue": 50000,
                                            "claimsLast3Years": 0
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wynik kalkulacji",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CalculationResult"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    }
                }
            }
        },
        "/claims": {
            "get": {
                "tags": [
                    "Claims"
                ],
                "summary": "Lista szkód",
                "operationId": "getClaims",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "clientId",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "policyId",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista szkód ze statystykami",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Claim"
                                            }
                                        },
                                        "total": {
                                            "type": "integer"
                                        },
                                        "stats": {
                                            "$ref": "#/components/schemas/ClaimStats"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Claims"
                ],
                "summary": "Zgłoś szkodę",
                "operationId": "createClaim",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClaimInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Szkoda zgłoszona"
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Policy": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "policyNumber": {
                        "type": "string",
                        "example": "OCPD/2024/001234"
                    },
                    "clientId": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "DRAFT",
                            "QUOTED",
                            "ACTIVE",
                            "EXPIRED",
                            "CANCELLED"
                        ]
                    },
                    "sumInsured": {
                        "type": "number",
                        "example": 500000
                    },
                    "totalPremium": {
                        "type": "number",
                        "example": 8800
                    },
                    "territorialScope": {
                        "type": "string",
                        "enum": [
                            "POLAND",
                            "EUROPE",
                            "WORLD"
                        ]
                    },
                    "validFrom": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "validTo": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "clauses": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PolicyClause"
                        }
                    }
                }
            },
            "PolicyInput": {
                "type": "object",
                "required": [
                    "clientId",
                    "sumInsured",
                    "territorialScope"
                ],
                "properties": {
                    "clientId": {
                        "type": "string"
                    },
                    "sumInsured": {
                        "type": "number"
                    },
                    "territorialScope": {
                        "type": "string",
                        "enum": [
                            "POLAND",
                            "EUROPE",
                            "WORLD"
                        ]
                    },
                    "clauses": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "PolicyClause": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "GROSS_NEGLIGENCE"
                    },
                    "sublimit": {
                        "type": "integer",
                        "example": 100
                    },
                    "premium": {
                        "type": "number",
                        "example": 704
                    }
                }
            },
            "PolicyStats": {
                "type": "object",
                "properties": {
                    "active": {
                        "type": "integer"
                    },
                    "expired": {
                        "type": "integer"
                    },
                    "totalPremium": {
                        "type": "number"
                    }
                }
            },
...
            "CalculationInput": {
                "type": "object",
                "required": [
                    "sumInsured",
                    "territorialScope"
                ],
                "properties": {
                    "sumInsured": {
                        "type": "number",
                        "description": "Suma ubezpieczenia w PLN",
                        "example": 300000
                    },
                    "territorialScope": {
                        "type": "string",
                        "enum": [
                            "POLAND",
                            "EUROPE",
                            "WORLD"
                        ]
                    },
                    "quickQuote": {
                        "type": "boolean",
                        "description": "True dla szybkiej wyceny"
                    },
                    "selectedClauses": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "GROSS_NEGLIGENCE",
                            "FRIDGE"
                        ]
                    },
                    "yearsInBusiness": {
                        "type": "integer"
                    },
                    "fleetSize": {
                        "type": "integer"
                    },
                    "apkData": {
                        "type": "object"
                    }
                }
            },
            "CalculationResult": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "enum": [
                            "quick",
                            "full"
                        ]
                    },
                    "estimatedPremium": {
                        "type": "number",
                        "description": "Dla quick quote"
                    },
                    "range": {
                        "type": "object",
                        "properties": {
                            "min": {
                                "type": "number"
                            },
                            "max": {
                                "type": "number"
                            }
                        }
                    },
                    "result": {
                        "type": "object",
                        "properties": {
                            "breakdown": {
                                "type": "object"
                            },
                            "riskLevel": {
                                "type": "string"
                            },
                            "isAutoApproved": {
                                "type": "boolean"
                            },
                            "referralReasons": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "Claim": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "claimNumber": {
                        "type": "string",
                        "example": "SZK/2024/00123"
                    },
                    "policyId": {
                        "type": "string"
                    },
                    "clientId": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "REPORTED",
                            "UNDER_REVIEW",
                            "APPROVED",
                            "REJECTED",
                            "PAID"
                        ]
                    },
                    "claimedAmount": {
                        "type": "number"
                    },
                    "paidAmount": {
                        "type": "number"
                    },
                    "description": {
                        "type": "string"
                    },
                    "incidentDate": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "ClaimInput": {
                "type": "object",
                "required": [
                    "policyId",
                    "clientId",
                    "description",
                    "claimedAmount"
                ],
                "properties": {
                    "policyId": {
                        "type": "string"
                    },
                    "clientId": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "claimedAmount": {
                        "type": "number"
                    },
                    "incidentDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "location": {
                        "type": "string"
                    }
                }
            },
            "ClaimStats": {
                "type": "object",
                "properties": {
                    "open": {
                        "type": "integer"
                    },
                    "paid": {
                        "type": "integer"
                    },
                    "totalClaimed": {
                        "type": "number"
                    },
                    "totalPaid": {
                        "type": "number"
                    }
                }
            }
        },
        "responses": {
            "BadRequest": {
                "description": "Nieprawidłowe dane wejściowe",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "Required field missing"
                                }
                            }
                        }
                    }
                }
            },
            "NotFound": {
                "description": "Zasób nie znaleziony",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "Resource not found"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}